Started Unit 2

+++ Unit One - Intro to Java/src/day4/ExampleThree.java
@@ -2,9 +2,10 @@ package day4;
 
 public class ExampleThree {
     public static void main(String[] args) {
-        int num = 4;
-        double x = 0;
+        int num = 4;            // no compile-time errors (no red squiggles)
+        int x = 0;
 
-        System.out.println(num / x);
+        System.out.println(num / x);    // run-time errors happen when the program is
+                                        // running. Java will throw an exception (ie. ArithmeticException)
     }
 }
